Move ReductiveRowSpec Generation out FieldCollectionHelper
