+++ generator/src/main/java/com/scottlogic/deg/generator/walker/reductive/FieldCollectionHelper.java
@@ -19,54 +19,22 @@ public class FieldCollectionHelper {
 
     private final GenerationConfig generationConfig;
     private final ConstraintReducer constraintReducer;
-    private final FieldSpecMerger fieldSpecMerger;
     private final FixFieldStrategy fixFieldStrategy;
     private final ReductiveDataGeneratorMonitor monitor;
 
     public FieldCollectionHelper(
         GenerationConfig config,
         ConstraintReducer constraintReducer,
-        FieldSpecMerger fieldSpecMerger,
         FixFieldStrategy fixFieldStrategy,
         ReductiveDataGeneratorMonitor monitor) {
-        this.fieldSpecMerger = fieldSpecMerger;
         this.fixFieldStrategy = fixFieldStrategy;
         this.generationConfig = config;
         this.constraintReducer = constraintReducer;
         this.monitor = monitor;
     }
 
-    //produce a stream of RowSpecs for each value in the permitted set of values for the field fixed on the last iteration
-    public Stream<RowSpec> createRowSpecsFromFixedValues(FieldCollection fieldCollection, ConstraintNode constraintNode) {
-        //create a row spec where every field is set to this.fixedFields & field=value
-        if (fieldCollection.getLastFixedField() == null) {
-            throw new UnsupportedOperationException("Field has not been fixed yet");
-        }
-
-        Map<Field, FieldSpec> fieldSpecsPerField = getFieldSpecsForAllFixedFieldsExceptLast(fieldCollection, constraintNode);
-
-        if (fieldSpecsPerField.values().stream().anyMatch(fieldSpec -> fieldSpec == FieldSpec.Empty)){
-            return Stream.empty();
-        }
-
-        FieldSpec fieldSpecForValuesInLastFixedField = fieldCollection.getLastFixedField().getFieldSpecForValues();
-        fieldSpecsPerField.put(fieldCollection.getLastFixedField().field, fieldSpecForValuesInLastFixedField);
-
-        RowSpec rowSpecWithAllValuesForLastFixedField = new ReductiveRowSpec(
-            fieldCollection.getFields(),
-            fieldSpecsPerField,
-            fieldCollection.getLastFixedField().field
-        );
-
-        this.monitor.rowSpecEmitted(
-            fieldCollection.getLastFixedField(),
-            fieldSpecForValuesInLastFixedField,
-            rowSpecWithAllValuesForLastFixedField);
-        return Stream.of(rowSpecWithAllValuesForLastFixedField);
-    }
-
     //work out the next field to fix and return a new FieldCollection with this field fixed
-    public FieldCollection fixNextField(FieldCollection fieldCollection, ReductiveConstraintNode rootNode) {
+    public FieldCollection fixNextFieldAndAddToCollection(FieldCollection fieldCollection, ReductiveConstraintNode rootNode) {
         Field fieldToFix = this.fixFieldStrategy.getNextFieldToFix(fieldCollection, rootNode);
 
         if (fieldToFix == null){
@@ -79,26 +47,6 @@ public class FieldCollectionHelper {
         FixedField field = createFixedFieldWithValues(fieldToFix, rootNode);
         return fieldCollection.with(field);
     }
-
-    //create a mapping of field->fieldspec for each fixed field - efficiency
-    private Map<Field, FieldSpec> getFieldSpecsForAllFixedFieldsExceptLast(FieldCollection fieldCollection, ConstraintNode constraintNode){
-        Map<Field, List<AtomicConstraint>> fieldToConstraints = constraintNode.getAtomicConstraints()
-            .stream()
-            .collect(Collectors.groupingBy(AtomicConstraint::getField));
-
-        return fieldCollection.getFixedFields().values()
-            .stream()
-            .collect(Collectors.toMap(
-                ff -> ff.field,
-                ff -> {
-                    FieldSpec fieldSpec = createFieldSpec(ff, fieldToConstraints.get(ff.field));
-                    return fieldSpec == null
-                        ? FieldSpec.Empty
-                        : fieldSpec;
-                }
-            ));
-    }
-
 
     //for the given field get a stream of possible values
     private FixedField createFixedFieldWithValues(Field field, ConstraintNode rootNode) {
@@ -120,18 +68,5 @@ public class FieldCollectionHelper {
         return new FixedField(field, values, rootConstraintsFieldSpec, this.monitor);
     }
 
-    //create a FieldSpec for a given FixedField and the atomic constraints we know about this field
-    private FieldSpec createFieldSpec(FixedField fixedField, Collection<AtomicConstraint> constraintsForField) {
-        FieldSpec fixedFieldSpec = fixedField.getFieldSpecForCurrentValue();
-        Optional<FieldSpec> constrainedFieldSpecOpt = this.constraintReducer.reduceConstraintsToFieldSpec(constraintsForField);
-
-        if (!constrainedFieldSpecOpt.isPresent()){
-            return null; //this shouldn't happen: caused by constraints for one of the fixed fields contradicting each other (issue in optimising and/or reducing) - see issue #250
-        }
-
-        return this.fieldSpecMerger
-            .merge(fixedFieldSpec, constrainedFieldSpecOpt.get())
-            .orElseThrow(() -> new UnsupportedOperationException("Contradiction? - " + fixedField.toString() + "\n" + constraintsForField.toString()));
-    }
 
 }
